feat: PortfolioDetail Service 추가

+++ src/main/java/com/example/shinhanserver/domain/pbinfo/service/dto/PBResponseDto.java
@@ -37,7 +37,7 @@ public class PBResponseDto {
     private List<String> portfolioNames;
     private List<Long> principals;
     private List<Long> returns;
-    private List<Long> cumulativeRORs;
+    private List<Double> cumulativeRORs;
     private List<String> durations;
 
 }
